requirements.txt
fastapi
uvicorn
streamlit
requests
python-dotenv
langchain
langchain-core
langchain-google-genai
crewai
pydantic
pillow